+++ uilayer-core/src/main/java/net/avh4/framework/uilayer/input/ClickReceiver.java
@@ -2,5 +2,5 @@ package net.avh4.framework.uilayer.input;
 
 public interface ClickReceiver {
 
-    void click(int x, int y);
+    void click(double x, double y);
 }
