+++ uilayer-core/src/main/java/net/avh4/framework/uilayer/mvc/Controller.java
@@ -1,7 +1,7 @@
 package net.avh4.framework.uilayer.mvc;
 
 public interface Controller<M> {
-    void click(M model, int x, int y);
+    void click(M model, double x, double y);
 
     void key(M Model, int keyCode, boolean shift);
 }
