+++ uilayer-core/src/main/java/net/avh4/framework/uilayer/scene/ClippedElementDelegate.java
@@ -2,5 +2,5 @@ package net.avh4.framework.uilayer.scene;
 
 public interface ClippedElementDelegate {
     void draw(GraphicsOperations g, FontMetricsService fm,
-              int clipLeftX, int clipTopY, int clipWidth, int clipHeight);
+              int clipLeftX, int clipTopY, double clipWidth, double clipHeight);
 }
