+++ uilayer-core/src/main/java/net/avh4/framework/uilayer/scene/GraphicsOperationsRenderer.java
@@ -1,5 +1,5 @@
 package net.avh4.framework.uilayer.scene;
 
 public interface GraphicsOperationsRenderer {
-    void render(int width, int height, GraphicsOperations g, FontMetricsService fm);
+    void render(double width, double height, GraphicsOperations g, FontMetricsService fm);
 }
