+++ uilayer-examples/src/main/java/net/avh4/demo/uilayer/ChooserDemo.java
@@ -27,7 +27,7 @@ public class ChooserDemo implements UI {
     }
 
     @Override
-    public void click(int x, int y) {
+    public void click(double x, double y) {
         if (x < 400) {
             chooseString();
         } else {
